01_jarvis/pyproject.toml
[tool.poetry]
name = "jarvis"
version = "0.1.0"
description = ""
authors = ["Ch Saqib <chsaqib11223@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
deepgram-sdk = "^3.2.7"
fastapi = "^0.111.0"
jinja2 = "^3.1.4"
groq = "^0.6.0"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
